moved: types/addr.go under crypto package
